relocate function
