+++ src/phoenix/server/api/schema.py
@@ -10,7 +10,7 @@ from strawberry import ID, UNSET
 from strawberry.types import Info
 from typing_extensions import Annotated
 
-from phoenix.config import DEFAULT_PROJECT_NAME, is_phoenix_server_instrumentation_enabled
+from phoenix.config import DEFAULT_PROJECT_NAME, is_server_instrumentation_enabled
 from phoenix.db import models
 from phoenix.pointcloud.clustering import Hdbscan
 from phoenix.server.api.context import Context
@@ -273,7 +273,7 @@ class Mutation(ExportEventsMutation):
 
 
 _extensions = []
-if is_phoenix_server_instrumentation_enabled():
+if is_server_instrumentation_enabled():
     from strawberry.extensions.tracing import OpenTelemetryExtension
 
     _extensions.append(OpenTelemetryExtension)
