+++ src/phoenix/server/main.py
@@ -17,8 +17,7 @@ from phoenix.config import (
     get_env_port,
     get_pids_path,
     get_working_dir,
-    initialize_opentelemetry_tracer_provider,
-    is_phoenix_server_instrumentation_enabled,
+    is_server_instrumentation_enabled,
 )
 from phoenix.core.model_schema_adapter import create_model_from_datasets
 from phoenix.db import get_printable_db_url
@@ -31,6 +30,7 @@ from phoenix.pointcloud.umap_parameters import (
     UMAPParameters,
 )
 from phoenix.server.app import create_app
+from phoenix.server.telemetry import initialize_opentelemetry_tracer_provider
 from phoenix.settings import Settings
 from phoenix.trace.fixtures import (
     TRACES_FIXTURES,
@@ -225,7 +225,7 @@ if __name__ == "__main__":
         initial_spans=fixture_spans,
         initial_evaluations=fixture_evals,
     )
-    if is_phoenix_server_instrumentation_enabled():
+    if is_server_instrumentation_enabled():
         from opentelemetry.instrumentation.starlette import StarletteInstrumentor
 
         initialize_opentelemetry_tracer_provider()
